requirements.txt
httpx>=0.25.0
pydantic>=2.5.0
pydantic-ai>=0.8.0
python-dotenv>=1.0.0
logfire>=0.8.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.15
typing-extensions>=4.8.0
openai>=1.3.0
langchain-community>=0.0.16
langchain-core>=0.1.14
langchain-openai>=0.0.2.post1
langgraph>=0.0.23
rich>=13.7.0
# Phụ thuộc mới cho API HTTP
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0 